sprint2/requirements.txt
pandas>=2.1
pymongo>=4.6
python-dotenv>=1.0
